+++ app/src/main/java/nextstep/payments/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import nextstep.payments.screens.card.NewCardScreen
+import nextstep.payments.screens.card.new.NewCardScreen
 import nextstep.payments.ui.theme.PaymentsTheme
 
 class MainActivity : ComponentActivity() {
